requirements.txt
feedparser>=6.0
python-telegram-bot>=20.0
requests>=2.28
rapidfuzz>=3.0
python-dotenv>=1.0
